+++ RELEASE-NOTES.md
@@ -32,6 +32,7 @@ as release notes for end user on a Recalbox upgrade.
   - As for USB devices, games are merged into your game list seamlessly
 - Add scraper progress bar
 - Add display by filename in gamelist option
+- Add go to game from search results and add search games from gamelist menu
 
 ### Improvements